some changes to the pledge model

+++ crowdfunding/projects/models.py
@@ -70,7 +70,7 @@ class Project(models.Model):
         super().save(**kwargs)
 
 class Pledge(models.Model):
-    amount = models.IntegerField()
+    amount = models.DecimalField(max_digits=20, decimal_places=2)
     comment = models.CharField(max_length=200)
     anonymous = models.BooleanField()
     project = models.ForeignKey('Project', on_delete=models.CASCADE, related_name='pledges')
@@ -79,6 +79,3 @@ class Pledge(models.Model):
         on_delete=models.CASCADE,
         related_name='pledges',
     )
-
-
-
